+++ src/npc/mod.rs
@@ -80,64 +80,29 @@ impl Plugin for NPCPlugin {
             .add_event::<aggression::StopChaseEvent>()
             .add_event::<aggression::DetectionModeEvent>()
             .add_event::<aggression::EngagePursuitEvent>()
-            .add_startup_system(spawn_characters)
-            .add_startup_system(spawn_aggresives_characters)
+            .add_systems(Startup, (spawn_characters, spawn_aggresives_characters))
             .add_systems(
+                FixedUpdate,
                 (
                     movement::just_walk.in_set(NPCSystems::Stroll),
                     idle::do_flexing
                         .in_set(NPCSystems::Idle)
                         .after(NPCSystems::Stroll),
                     movement::follow.in_set(NPCSystems::Follow),
-                )
-                    .in_base_set(CoreSet::Update)
-                    .in_schedule(CoreSchedule::FixedUpdate),
-            )
-            // .add_system(aggression::add_pursuit_urge)
-            // .add_system(aggression::remove_pursuit_urge)
-            .add_system(
-                aggression::add_detection_aura
-                    .before(NPCSystems::FindTargets)
-                    .in_base_set(CoreSet::Update)
-                    .in_schedule(CoreSchedule::FixedUpdate),
-            )
-            .add_system(
-                aggression::threat_detection
-                    .in_set(NPCSystems::FindTargets)
-                    .in_base_set(CoreSet::Update)
-                    .in_schedule(CoreSchedule::FixedUpdate),
-            )
-            .add_system(
-                aggression::add_pursuit_urge
-                    .before(NPCSystems::Chase)
-                    .after(NPCSystems::FindTargets)
-                    .in_base_set(CoreSet::Update)
-                    .in_schedule(CoreSchedule::FixedUpdate),
-            )
-            .add_system(
-                movement::pursue
-                    .in_set(NPCSystems::Chase)
-                    .after(NPCSystems::FindTargets)
-                    .in_base_set(CoreSet::Update)
-                    .in_schedule(CoreSchedule::FixedUpdate),
-            )
-            .add_system(
-                aggression::remove_pursuit_urge
-                    .in_set(NPCSystems::StopChase)
-                    .after(NPCSystems::Chase)
-                    .in_base_set(CoreSet::Update)
-                    .in_schedule(CoreSchedule::FixedUpdate),
-            )
-            .add_system(
-                aggression::fair_play_wait
-                    .after(NPCSystems::StopChase)
-                    .in_base_set(CoreSet::Update)
-                    .in_schedule(CoreSchedule::FixedUpdate),
-            )
-            .add_system(
-                aggression::add_detection_aura
-                    .after(NPCSystems::StopChase)
-                    .in_schedule(CoreSchedule::FixedUpdate),
+                    aggression::add_detection_aura.before(NPCSystems::FindTargets),
+                    aggression::threat_detection.in_set(NPCSystems::FindTargets),
+                    aggression::add_pursuit_urge
+                        .before(NPCSystems::Chase)
+                        .after(NPCSystems::FindTargets),
+                    movement::pursue
+                        .in_set(NPCSystems::Chase)
+                        .after(NPCSystems::FindTargets),
+                    aggression::remove_pursuit_urge
+                        .in_set(NPCSystems::StopChase)
+                        .after(NPCSystems::Chase),
+                    aggression::fair_play_wait.after(NPCSystems::StopChase),
+                    aggression::add_detection_aura.after(NPCSystems::StopChase),
+                ),
             );
     }
 }
@@ -170,7 +135,7 @@ fn spawn_characters(mut commands: Commands, fabien: Res<FabienSheet>) {
                 speed: Speed::default(),
                 velocity: Velocity {
                     linvel: Vect::ZERO,
-                    angvel: 0.0,
+                    angvel: 0.,
                 },
             },
             CombatBundle {
@@ -186,7 +151,7 @@ fn spawn_characters(mut commands: Commands, fabien: Res<FabienSheet>) {
         .with_children(|parent| {
             parent.spawn((
                 Collider::cuboid(CHAR_HITBOX_WIDTH, CHAR_HITBOX_HEIGHT),
-                Transform::from_xyz(0.0, CHAR_HITBOX_Y_OFFSET, 0.0),
+                Transform::from_xyz(0., CHAR_HITBOX_Y_OFFSET, 0.),
                 CharacterHitbox,
             ));
         });
@@ -215,7 +180,7 @@ fn spawn_characters(mut commands: Commands, fabien: Res<FabienSheet>) {
                 speed: Speed::default(),
                 velocity: Velocity {
                     linvel: Vect::ZERO,
-                    angvel: 0.0,
+                    angvel: 0.,
                 },
             },
             CombatBundle {
@@ -231,7 +196,7 @@ fn spawn_characters(mut commands: Commands, fabien: Res<FabienSheet>) {
         .with_children(|parent| {
             parent.spawn((
                 Collider::cuboid(CHAR_HITBOX_WIDTH, CHAR_HITBOX_HEIGHT),
-                Transform::from_xyz(0.0, CHAR_HITBOX_Y_OFFSET, 0.0),
+                Transform::from_xyz(0., CHAR_HITBOX_Y_OFFSET, 0.),
                 CharacterHitbox,
             ));
         });
@@ -266,7 +231,7 @@ fn spawn_aggresives_characters(mut commands: Commands, fabien: Res<FabienSheet>)
                 speed: Speed(NPC_SPEED_LEADER),
                 velocity: Velocity {
                     linvel: Vect::ZERO,
-                    angvel: 0.0,
+                    angvel: 0.,
                 },
             },
             CombatBundle {
@@ -288,7 +253,7 @@ fn spawn_aggresives_characters(mut commands: Commands, fabien: Res<FabienSheet>)
         .with_children(|parent| {
             parent.spawn((
                 Collider::cuboid(CHAR_HITBOX_WIDTH, CHAR_HITBOX_HEIGHT),
-                Transform::from_xyz(0.0, CHAR_HITBOX_Y_OFFSET, 0.0),
+                Transform::from_xyz(0., CHAR_HITBOX_Y_OFFSET, 0.),
                 CharacterHitbox,
             ));
 
@@ -334,7 +299,7 @@ fn spawn_aggresives_characters(mut commands: Commands, fabien: Res<FabienSheet>)
                     speed: Speed(NPC_SPEED),
                     velocity: Velocity {
                         linvel: Vect::ZERO,
-                        angvel: 0.0,
+                        angvel: 0.,
                     },
                 },
                 CombatBundle {
@@ -356,7 +321,7 @@ fn spawn_aggresives_characters(mut commands: Commands, fabien: Res<FabienSheet>)
             .with_children(|parent| {
                 parent.spawn((
                     Collider::cuboid(CHAR_HITBOX_WIDTH, CHAR_HITBOX_HEIGHT),
-                    Transform::from_xyz(0.0, CHAR_HITBOX_Y_OFFSET, 0.0),
+                    Transform::from_xyz(0., CHAR_HITBOX_Y_OFFSET, 0.),
                     CharacterHitbox,
                 ));